refactored lock list
