+++ app/package.json
@@ -10,6 +10,7 @@
         "@ionic/react-router": "^4.11.0",
         "@material-ui/core": "^4.9.4",
         "@material-ui/icons": "^4.9.1",
+        "@polkadot/api": "^1.19.1",
         "@polkadot/util-crypto": "^2.9.1",
         "@types/bigi": "^1.4.2",
         "@types/bitcore-lib": "^0.15.1",
